native boards: Split SDL events handling in top and bottom
Split the SDL event handling code in a top and bottom
to enable building it with embedded libCs.
