+++ boards/posix/common/sdl/sdl_events.c
@@ -8,53 +8,19 @@
 #include "soc.h"
 #include <zephyr/arch/posix/posix_trace.h>
 #include <zephyr/kernel.h>
-
-#include <SDL.h>
-
-static void sdl_handle_window_event(const SDL_Event *event)
-{
-	SDL_Window *window;
-	SDL_Renderer *renderer;
-
-	switch (event->window.event) {
-	case SDL_WINDOWEVENT_EXPOSED:
-
-		window = SDL_GetWindowFromID(event->window.windowID);
-		if (window == NULL) {
-			return;
-		}
-
-		renderer = SDL_GetRenderer(window);
-		if (renderer == NULL) {
-			return;
-		}
-		SDL_RenderPresent(renderer);
-		break;
-	default:
-		break;
-	}
-}
+#include "sdl_events_bottom.h"
 
 static void sdl_handle_events(void *p1, void *p2, void *p3)
 {
-	SDL_Event event;
-
 	ARG_UNUSED(p1);
 	ARG_UNUSED(p2);
 	ARG_UNUSED(p3);
 
 	for (;;) {
-		while (SDL_PollEvent(&event)) {
-			switch (event.type) {
-			case SDL_WINDOWEVENT:
-				sdl_handle_window_event(&event);
-				break;
-			case SDL_QUIT:
-				posix_exit(0);
-				break;
-			default:
-				break;
-			}
+		int rc = sdl_handle_pending_events();
+
+		if (rc != 0) {
+			posix_exit(0);
 		}
 
 		k_msleep(CONFIG_SDL_THREAD_INTERVAL);
@@ -63,15 +29,14 @@ static void sdl_handle_events(void *p1, void *p2, void *p3)
 
 static void sdl_init(void)
 {
-	if (SDL_Init(SDL_INIT_VIDEO) != 0) {
-		posix_print_error_and_exit("Error on SDL_Init (%s)\n",
-				SDL_GetError());
+	if (sdl_init_video() != 0) {
+		posix_print_error_and_exit("Error on SDL_Init (%s)\n", sdl_get_error());
 	}
 }
 
 static void sdl_cleanup(void)
 {
-	SDL_Quit();
+	sdl_quit();
 }
 
 NATIVE_TASK(sdl_init, PRE_BOOT_2, 1);
